requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.5
httpx==0.25.2
google-generativeai==0.3.1
